Schedule: Wire up print button
This is the best I think we're going to be able to do in the time
we have left. Something in materialize doesn't play nice with
fullcalendar such that whenever the window gets rendered in print
mode events aren't moved even though the calendar is, which makes
them off (sometimes by minutes, sometimes by hours). It's not even
linear.

To compromise, I'm rendering the print in a separate page that only
includes the fullcalendar stylesheet and a few other tweaks. This
means that compared to the rest of the site it looks fugly and there
is no color, but given that most students probably only have access
to a B/W printer I think this is probably fine. They're not looking
at this outside of bringing it with them when they go to schedule
classes anyways.